Migration changed

+++ db/migrate/20130320070224_change_data_type_for_boolean.rb
@@ -1,6 +1,6 @@
 class ChangeDataTypeForBoolean < ActiveRecord::Migration
   def up
-  change_column :evaluation_frequencies, :format, :float
+  change_column :evaluation_frequencies, :format, :decimal
   end
 
   def down
